refactor: move remaining provider onboarding into extensions
