+++ extensions/zai/index.ts
@@ -20,17 +20,13 @@ import {
 } from "../../src/commands/auth-choice.api-key.js";
 import { ensureApiKeyFromOptionEnvOrPrompt } from "../../src/commands/auth-choice.apply-helpers.js";
 import { buildApiKeyCredential } from "../../src/commands/onboard-auth.credentials.js";
-import {
-  applyAuthProfileConfig,
-  applyZaiConfig,
-  applyZaiProviderConfig,
-  ZAI_DEFAULT_MODEL_REF,
-} from "../../src/commands/onboard-auth.js";
+import { applyAuthProfileConfig } from "../../src/commands/onboard-auth.js";
 import type { SecretInput } from "../../src/config/types.secrets.js";
 import { resolveRequiredHomeDir } from "../../src/infra/home-dir.js";
 import { fetchZaiUsage } from "../../src/infra/provider-usage.fetch.js";
 import { normalizeOptionalSecretInput } from "../../src/utils/normalize-secret-input.js";
 import { detectZaiEndpoint, type ZaiEndpointId } from "./detect.js";
+import { applyZaiConfig, applyZaiProviderConfig, ZAI_DEFAULT_MODEL_REF } from "./onboard.js";
 
 const PROVIDER_ID = "zai";
 const GLM5_MODEL_ID = "glm-5";
