+++ src/commands/onboard-auth.ts
@@ -59,7 +59,6 @@ export {
 } from "./onboard-auth.config-opencode-go.js";
 export {
   CLOUDFLARE_AI_GATEWAY_DEFAULT_MODEL_REF,
-  KILOCODE_DEFAULT_MODEL_REF,
   LITELLM_DEFAULT_MODEL_REF,
   OPENROUTER_DEFAULT_MODEL_REF,
   setOpenaiApiKey,
@@ -90,15 +89,16 @@ export {
   writeOAuthCredentials,
   VERCEL_AI_GATEWAY_DEFAULT_MODEL_REF,
   XIAOMI_DEFAULT_MODEL_REF,
-  ZAI_DEFAULT_MODEL_REF,
-  MODELSTUDIO_DEFAULT_MODEL_REF,
 } from "./onboard-auth.credentials.js";
 export { HUGGINGFACE_DEFAULT_MODEL_REF } from "../../extensions/huggingface/onboard.js";
+export { KILOCODE_DEFAULT_MODEL_REF } from "../../extensions/kilocode/onboard.js";
 export { MISTRAL_DEFAULT_MODEL_REF } from "../../extensions/mistral/onboard.js";
+export { MODELSTUDIO_DEFAULT_MODEL_REF } from "../../extensions/modelstudio/onboard.js";
 export { SYNTHETIC_DEFAULT_MODEL_REF } from "../../extensions/synthetic/onboard.js";
 export { TOGETHER_DEFAULT_MODEL_REF } from "../../extensions/together/onboard.js";
 export { VENICE_DEFAULT_MODEL_REF } from "../../extensions/venice/onboard.js";
 export { XAI_DEFAULT_MODEL_REF } from "../../extensions/xai/onboard.js";
+export { ZAI_DEFAULT_MODEL_REF } from "../../extensions/zai/onboard.js";
 export {
   buildKilocodeModelDefinition,
   buildMinimaxApiModelDefinition,
